+++ components/table/storage/block_manager.cpp
@@ -14,7 +14,7 @@ namespace components::table::storage {
         : buffer_manager(buffer_manager)
         , block_alloc_size_(block_alloc_size) {}
 
-    std::shared_ptr<block_handle_t> block_manager_t::register_block(uint32_t block_id) {
+    std::shared_ptr<block_handle_t> block_manager_t::register_block(uint64_t block_id) {
         std::lock_guard lock(blocks_lock_);
         auto entry = blocks_.find(block_id);
         if (entry != blocks_.end()) {
@@ -28,7 +28,7 @@ namespace components::table::storage {
         return result;
     }
 
-    void block_manager_t::unregister_block(uint32_t id) {
+    void block_manager_t::unregister_block(uint64_t id) {
         assert(id < MAXIMUM_BLOCK);
         std::lock_guard lock(blocks_lock_);
         blocks_.erase(id);
