+++ components/table/storage/block_manager.hpp
@@ -19,20 +19,20 @@ namespace components::table::storage {
 
         buffer_manager_t& buffer_manager;
 
-        virtual std::unique_ptr<block_t> convert_block(uint32_t block_id, file_buffer_t& source_buffer) = 0;
-        virtual std::unique_ptr<block_t> create_block(uint32_t block_id, file_buffer_t* source_buffer) = 0;
+        virtual std::unique_ptr<block_t> convert_block(uint64_t block_id, file_buffer_t& source_buffer) = 0;
+        virtual std::unique_ptr<block_t> create_block(uint64_t block_id, file_buffer_t* source_buffer) = 0;
 
-        virtual uint32_t free_block_id() = 0;
-        virtual uint32_t peek_free_block_id() = 0;
+        virtual uint64_t free_block_id() = 0;
+        virtual uint64_t peek_free_block_id() = 0;
         virtual bool is_root_block(meta_block_pointer_t root) = 0;
-        virtual void mark_as_free(uint32_t block_id) = 0;
-        virtual void mark_as_used(uint32_t block_id) = 0;
-        virtual void mark_as_modified(uint32_t block_id) = 0;
-        virtual void increase_block_ref_count(uint32_t block_id) = 0;
+        virtual void mark_as_free(uint64_t block_id) = 0;
+        virtual void mark_as_used(uint64_t block_id) = 0;
+        virtual void mark_as_modified(uint64_t block_id) = 0;
+        virtual void increase_block_ref_count(uint64_t block_id) = 0;
         virtual uint64_t meta_block() = 0;
         virtual void read(block_t& block) = 0;
-        virtual void read_blocks(file_buffer_t& buffer, uint32_t start_block, uint64_t block_count) = 0;
-        virtual void write(file_buffer_t& block, uint32_t block_id) = 0;
+        virtual void read_blocks(file_buffer_t& buffer, uint64_t start_block, uint64_t block_count) = 0;
+        virtual void write(file_buffer_t& block, uint64_t block_id) = 0;
         void write(block_t& block) { write(block, block.id); }
 
         virtual uint64_t total_blocks() = 0;
@@ -42,10 +42,10 @@ namespace components::table::storage {
         virtual void file_sync() = 0;
         virtual void truncate();
 
-        std::shared_ptr<block_handle_t> register_block(uint32_t block_id);
+        std::shared_ptr<block_handle_t> register_block(uint64_t block_id);
 
         void unregister_block(block_handle_t& block);
-        void unregister_block(uint32_t id);
+        void unregister_block(uint64_t id);
 
         uint64_t block_allocation_size() const { return block_alloc_size_; }
         uint64_t block_size() const { return block_alloc_size_ - DEFAULT_BLOCK_HEADER_SIZE; }
@@ -58,7 +58,7 @@ namespace components::table::storage {
 
     private:
         std::mutex blocks_lock_;
-        std::unordered_map<uint32_t, std::weak_ptr<block_handle_t>> blocks_;
+        std::unordered_map<uint64_t, std::weak_ptr<block_handle_t>> blocks_;
         uint64_t block_alloc_size_;
     };
